+++ lib/pay/braintree/charge.rb
@@ -5,6 +5,23 @@ module Pay
 
       delegate :processor_id, to: :pay_charge
 
+      def self.sync(charge_id, object: nil, try: 0, retries: 1)
+        object ||= Pay.braintree_gateway.transaction.find(charge_id)
+
+        pay_customer = Pay::Customer.find_by(processor: :braintree, processor_id: object.customer_details.id)
+        return unless pay_customer
+
+        pay_customer.save_transaction(object)
+      rescue ActiveRecord::RecordInvalid, ActiveRecord::RecordNotUnique
+        try += 1
+        if try <= retries
+          sleep 0.1
+          retry
+        else
+          raise
+        end
+      end
+
       def initialize(pay_charge)
         @pay_charge = pay_charge
       end